ipopt: update urls to use github instead of coin-or.org, add version 3.14.5, add maintainer (#30207)
* ipopt: add goxberry as maintainer

This commit adds 'goxberry' (me, Geoff Oxberry) as a maintainer of the
Ipopt Spack package.

* ipopt: use github url instead of coin-or.org url

This commit changes the package URL for Ipopt from one containing
`coin-or.org` to one containing `github.com`. The rationale for
using `github.com` is as follows:

- The COIN-OR webpage now directs users interested in Ipopt source to
  GitHub.

- Ipopt used to have a COIN-OR project homepage actually hosted on
  coin-or.org using an SVN-Trac web page. A link to this project
  homepage no longer appears within the "Projects" section of
  COIN-OR's website.

- COIN-OR issued a 2021-12-15 post on the News section of its web site
  (see https://www.coin-or.org/news/) that discusses the impact that
  lack of financial support has on COIN-OR software maintenance. It
  seems reasonable to suspect that the GitHub project is likely to
  outlast the COIN-OR web site.

The sha256 hashes for ipopt@:3.12 downloaded from GitHub differ from
the corresponding COIN-OR versions, so these hashes are also updated.

* ipopt 3.14.5: add new version

This commit adds the latest version of Ipopt, 3.14.5, to the Ipopt
Spack package.